Move more stuff to separate classes and adapt + cleanup includes
