+++ arangod/Replication2/Storage/RocksDB/LogIterator.h
@@ -24,7 +24,7 @@
 #pragma once
 
 #include "Basics/RocksDBUtils.h"
-#include "Replication2/ReplicatedLog/LogEntries.h"
+#include "Replication2/ReplicatedLog/PersistingLogEntry.h"
 #include "RocksDBEngine/RocksDBKey.h"
 #include "RocksDBEngine/RocksDBKeyBounds.h"
 #include "RocksDBEngine/RocksDBValue.h"
