+++ arangod/Replication2/Streams/Streams.h
@@ -23,7 +23,7 @@
 #pragma once
 
 #include "Replication2/ReplicatedLog/LogCommon.h"
-#include "Replication2/ReplicatedLog/LogEntries.h"
+#include "Replication2/ReplicatedLog/TypedLogIterator.h"
 #include "Replication2/Streams/StreamSpecification.h"
 
 namespace arangodb {
